[UPDATE] id값 타입 변경

+++ src/main/java/com/dsmpear/main/user_backend_v2/controller/NoticeController.java
@@ -23,7 +23,7 @@ public class NoticeController {
     }
 
     @GetMapping("/{noticeId}")
-    public NoticeContentResponse getNoticeContent(@PathVariable Integer noticeId) {
+    public NoticeContentResponse getNoticeContent(@PathVariable Long noticeId) {
         return noticeService.getNoticeContent(noticeId);
     }
